+++ firmware/src/hal/x86/test/run-all.sh
@@ -2,7 +2,7 @@
 
 if [[ $1 == "exec" ]]; then
     BASEDIR=$(realpath $(dirname $0))
-    TESTDIRS="hmac_sha256"
+    TESTDIRS="bip32_path hmac_sha256"
     for d in $TESTDIRS; do
         echo "******************************"
         echo "Testing $d..."
